fix: Rename tests
