+++ tests/test_all.py
@@ -1,6 +1,8 @@
 from graphcompress.readers import Reader_NT, Reader_NTgz, Reader_CGgz
-from graphcompress.parsers import WikidataParser, CGParser
-from graphcompress.builders import Builder
+from graphcompress.parsers import WikidataParser, CGParser, IGParser
+from graphcompress.builders.compress_graph import CGBuilder
+from graphcompress.builders.index_graph import IGBuilder
+
 import os
 import unittest
 
@@ -31,17 +33,31 @@ class GraphcompressTest(unittest.TestCase):
         os.rmdir(os.path.join(ACTUAL_PATH,'tmp'))
         pass
 
-    def test_all(self):
+    def test_compressed_format(self):
         r = Reader_NT(self.test_file)
         p = WikidataParser()
-        b = Builder(r,p,self.partitions_dir,self.output_file)
+        b = CGBuilder(r,p,self.partitions_dir,self.output_file)
 
         b.make_partitions()
         b.merge_partitions()
 
         rc = Reader_CGgz(self.output_file)
         pc = CGParser()
-        self.assertEqual(pc.parse_line(next(rc.readline())), (1, [[132, 2, 3]]))        
+
+        self.assertEqual(pc.parse_line(next(rc.readline())), (1, [[132, 2, 3]]))
+
+    def test_indexed_format(self):
+        r = Reader_NT(self.test_file)
+        p = WikidataParser()
+        b = IGBuilder(r,p,self.partitions_dir,self.output_file)
+
+        b.make_partitions()
+        b.merge_partitions()
+
+        pc = IGParser()
+        rc = Reader_CGgz(self.output_file)
+
+        self.assertEqual(pc.parse_line(next(rc.readline())), (1, [0,1]))
